Move all gateway messages into tunnel
